Fix merge conflicts and add CI quality gates
- Resolved merge conflicts in behav_assessment_model.py, interview_system.py, question.py, user_activity_service.py
- Fixed datetime import issues in listening1.py
- Added CI/CD pipeline with lint, typecheck, and test gates for dev->QA promotion
- Fixed test imports and assertions in test_sm2.py
- Added Pod 3 routes (interview, practice, resume) to main router
- All Pod 3 routes are auth-protected with get_current_user_id dependency

+++ app/services/confidence_analyzer.py
@@ -76,4 +76,4 @@ def compute_confidence(features: dict) -> int:
     weight_sum += 1
 
     confidence = score / weight_sum
-    return int(round(confidence))
+    return round(confidence)
